Index clients on ProxiedPlayer instance instead of UUID.
This fixes a bug caused by a huge gotcha in the way BungeeCord handles connections, as there can be two connections with the same UUID present. The previous implementation would overwrite the existing IRC client once the second Minecraft client was connected, then delete it again when the first Minecraft client was disconnected. This caused leaking of IRC client instances, a whole lot of NPEs, and chat messages not being relayed to IRC.